Added '--forget-inc-paths' to the assembler command line.
The file 'c64.inc' is already part of the cc65 suite and the ca65
assembler seems to prefer the system's version to the one in the 'inc'
directory. With the added option the assembler will forget it's include
paths and use the includes in the directories supplied on the
commandline.